requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
python-dotenv==0.19.0
selenium==4.16.0
webdriver-manager==4.0.2
requests==2.32.3
beautifulsoup4==4.9.3
lxml==4.6.3
python-multipart==0.0.5
aiohttp==3.8.1
groq==0.3.0
langchain==0.0.350
openai==1.3.5
httpx==0.25.1
langchain-groq==0.3.2 